requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==2.4.1
blinker==1.9.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.9
distro==1.9.0
Django==5.1.5
dronekit==2.9.2
dronekit-sitl==3.3.0
executing==2.0.1
filelock==3.16.1
Flask==3.1.0
future==0.18.3
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.6
ipykernel==6.28.0
ipython==8.19.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.9.0
jupyter_client==8.6.0
jupyter_core==5.7.0
lxml==4.9.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
MAVProxy==1.8.67
monotonic==1.6
nest-asyncio==1.5.8
numpy==1.26.1
openai==1.72.0
packaging==23.2
parso==0.8.3
pefile==2023.2.7
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.6
pure-eval==0.2.2
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.17.2
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
pymavlink==2.4.40
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.1.0
pytube==15.0.0
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==25.1.2
requests==2.31.0
setuptools==69.0.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
stack-data==0.6.3
style==1.1.0
tk==0.1.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.2
update==0.0.1
urllib3==2.1.0
virtualenv==20.27.1
virtualenvwrapper-win==1.2.7
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.41.3
